requirements.txt
aiohttp==3.8.4
tensorflow==2.8.0
discord.py_self==2.0.0
numpy==1.22.2
Pillow==10.0.0
aiosqlite==0.19.0
discord==2.3.1
